[Gold IV] Title: 타임머신, Time: 12 ms, Memory: 2444 KB -BaekjoonHub

+++ 백준/Gold/11657. 타임머신/타임머신.cc
@@ -6,15 +6,15 @@ using namespace std;
 
 long long n, m, a, b, c;
 vector<long long> dist(501, 1e9 + 7);
-vector<vector<long long>> edges(6000, vector<long long>(3));
+vector<vector<int>> edges(6000, vector<int>(3));
 
 bool bf(int start) {
 	dist[start] = 0;
 	for (int i = 0; i < n; i++) {
 		for (int j = 0; j < m; j++) {
-			long long cur_node = edges[j][0];
-			long long next_node = edges[j][1];
-			long long cost = edges[j][2];
+			int cur_node = edges[j][0];
+			int next_node = edges[j][1];
+			int cost = edges[j][2];
 			if (dist[cur_node] != 1e9+7 && dist[next_node] > dist[cur_node] + cost) {
 				dist[next_node] = dist[cur_node] + cost;
 				if (i == n - 1) {
